structure adjusted
