+++ src/ui/mod.rs
@@ -1,0 +1,6 @@
+/// jpdict/src/ui/mod.rs
+
+mod constants;
+mod app;
+
+pub use app::DictionaryApp;
